# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Shoppable Video App API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Video {
  id: number;
  title: string;
  /** URL of the Instagram post (for embedding) */
  instagramUrl: string;
  /** Direct video file URL (if uploaded) */
  videoUrl: string | null;
  thumbnailUrl: string | null;
  /** Instagram profile/page URL that clicking the video should link to */
  instagramPageUrl: string | null;
  description: string | null;
  tags: string[];
  createdAt: string;
  updatedAt: string;
}

export interface CreateVideoBody {
  title: string;
  instagramUrl: string;
  videoUrl?: string | null;
  thumbnailUrl?: string | null;
  instagramPageUrl?: string | null;
  description?: string | null;
  tags?: string[];
}

export interface UpdateVideoBody {
  title?: string;
  instagramUrl?: string;
  videoUrl?: string | null;
  thumbnailUrl?: string | null;
  instagramPageUrl?: string | null;
  description?: string | null;
  tags?: string[];
}

export interface VideoListResponse {
  videos: Video[];
  total: number;
  page: number;
  limit: number;
}

/**
 * product_page shows 3-4 videos; homepage shows 6-7 videos
 */
export type WidgetType = (typeof WidgetType)[keyof typeof WidgetType];

export const WidgetType = {
  product_page: "product_page",
  homepage: "homepage",
} as const;

export interface Widget {
  id: number;
  name: string;
  /** product_page shows 3-4 videos; homepage shows 6-7 videos */
  type: WidgetType;
  /** Shopify product ID or page handle where this widget appears */
  shopifyContext: string | null;
  isActive: boolean;
  createdAt: string;
  updatedAt: string;
}

export type WidgetWithVideos = Widget & {
  videos: Video[];
};

export type CreateWidgetBodyType =
  (typeof CreateWidgetBodyType)[keyof typeof CreateWidgetBodyType];

export const CreateWidgetBodyType = {
  product_page: "product_page",
  homepage: "homepage",
} as const;

export interface CreateWidgetBody {
  name: string;
  type: CreateWidgetBodyType;
  shopifyContext?: string | null;
  isActive?: boolean;
}

export type UpdateWidgetBodyType =
  (typeof UpdateWidgetBodyType)[keyof typeof UpdateWidgetBodyType];

export const UpdateWidgetBodyType = {
  product_page: "product_page",
  homepage: "homepage",
} as const;

export interface UpdateWidgetBody {
  name?: string;
  type?: UpdateWidgetBodyType;
  shopifyContext?: string | null;
  isActive?: boolean;
}

export interface SetWidgetVideosBody {
  /** Ordered list of video IDs to assign to this widget */
  videoIds: number[];
}

export interface WidgetListResponse {
  widgets: Widget[];
  total: number;
}

export interface Stats {
  totalVideos: number;
  totalWidgets: number;
  productPageWidgets: number;
  homepageWidgets: number;
  activeWidgets: number;
  recentVideos: Video[];
}

export type EmbedDataWidgetType =
  (typeof EmbedDataWidgetType)[keyof typeof EmbedDataWidgetType];

export const EmbedDataWidgetType = {
  product_page: "product_page",
  homepage: "homepage",
} as const;

export type EmbedDataVideosItem = {
  id: number;
  title: string;
  videoUrl: string | null;
  thumbnailUrl: string | null;
  instagramUrl: string;
  instagramPageUrl: string | null;
};

export interface EmbedData {
  widgetId: number;
  widgetType: EmbedDataWidgetType;
  videos: EmbedDataVideosItem[];
}

export type ListVideosParams = {
  page?: number;
  limit?: number;
};